+++ src/grammar/ast.ts
@@ -81,13 +81,6 @@ export type ASTNull = {
     ref: ASTRef;
 };
 
-export type ASTLvalueRef = {
-    kind: "lvalue_ref";
-    id: number;
-    name: string;
-    ref: ASTRef;
-};
-
 //
 // Types
 //
@@ -166,7 +159,7 @@ export type ASTOpField = {
     kind: "op_field";
     id: number;
     src: ASTExpression;
-    name: string;
+    name: ASTId;
     ref: ASTRef;
 };
 
@@ -439,7 +432,7 @@ export type ASTStatementExpression = {
 export type ASTStatementAssign = {
     kind: "statement_assign";
     id: number;
-    path: ASTLvalueRef[];
+    path: ASTExpression;
     expression: ASTExpression;
     ref: ASTRef;
 };
@@ -458,7 +451,7 @@ export type ASTStatementAugmentedAssign = {
     kind: "statement_augmentedassign";
     id: number;
     op: ASTAugmentedAssignOperation;
-    path: ASTLvalueRef[];
+    path: ASTExpression;
     expression: ASTExpression;
     ref: ASTRef;
 };
@@ -518,7 +511,7 @@ export type ASTStatementForEach = {
     id: number;
     keyName: string;
     valueName: string;
-    map: ASTLvalueRef[];
+    map: ASTExpression;
     statements: ASTStatement[];
     ref: ASTRef;
 };
@@ -570,7 +563,6 @@ export type ASTNode =
     | ASTStatementTryCatch
     | ASTStatementForEach
     | ASTReceive
-    | ASTLvalueRef
     | ASTString
     | ASTTrait
     | ASTProgramImport
@@ -587,11 +579,29 @@ export type ASTExpression =
     | ASTOpCallStatic
     | ASTOpNew
     | ASTNull
-    | ASTLvalueRef
     | ASTInitOf
     | ASTString
     | ASTConditional;
 export type ASTType = ASTPrimitive | ASTStruct | ASTContract | ASTTrait;
+
+/**
+ * Check if input expression is a 'path expression',
+ * i.e. an identifier or a sequence of field accesses starting from an identifier.
+ * @param path A path expression to check.
+ * @returns An array of identifiers or null if the input expression is not a path expression.
+ */
+export function tryExtractPath(path: ASTExpression): ASTId[] | null {
+    switch (path.kind) {
+        case "id":
+            return [path];
+        case "op_field": {
+            const p = tryExtractPath(path.src);
+            return p ? [...p, path.name] : null;
+        }
+        default:
+            return null;
+    }
+}
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type DistributiveOmit<T, K extends keyof any> = T extends any
@@ -713,15 +723,11 @@ export function traverse(node: ASTNode, callback: (node: ASTNode) => void) {
         traverse(node.expression, callback);
     }
     if (node.kind === "statement_assign") {
-        for (const e of node.path) {
-            traverse(e, callback);
-        }
+        traverse(node.path, callback);
         traverse(node.expression, callback);
     }
     if (node.kind === "statement_augmentedassign") {
-        for (const e of node.path) {
-            traverse(e, callback);
-        }
+        traverse(node.path, callback);
         traverse(node.expression, callback);
     }
     if (node.kind === "statement_condition") {